% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_ss_records.R
\name{analyse_ss_records}
\alias{analyse_ss_records}
\title{Analyse stop and search records}
\usage{
analyse_ss_records(data, csvs = FALSE)
}
\arguments{
\item{data}{A data frame containing the stop records to analyse}

\item{csvs}{If TRUE, writes a separate csv for each ethnicity}
}
\value{
A list of data frames containing the frequency and rate of stops of
each ethnicity in each local authority in data, and a comparison of this rate
to the rate for the White population in the form of odds and risk ratios. If
csvs = T, also writes a separate csv file for each ethnicity in data.
}
\description{
Calculates the number of stop and searches of each ethnicity in each local
authority present in data, and expresses these as a function of the size of
the respective population of each ethnicity. Also calculates the odds ratio
and risk ratio of the stop rate of each minority ethnicity compared to the
stop rate of the White ethnic category. Minority ethnicities have been
aggregated into broad categories. Mixed categories are defined based on the
minority component (e.g., Mixed White and Black African is categorised as
Mixed Black). White ethnicity is composed of 'White - Any other White background',
White - English/Welsh/Scottish/Northern Irish/British', and 'White - Irish'.
}
\examples{
outputs <- analyse(data)

# Individual data frames can then be acquired by indexing the list, e.g.

stops_Black <- outputs[["Black"]]

}
