% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ss_data.R
\name{extract_ss_data}
\alias{extract_ss_data}
\title{Extract stop and search records}
\usage{
extract_ss_data(
  coord_list = coords,
  num_months_backwards = 12,
  oldest_month = NULL,
  oldest_year = NULL,
  most_recent_month = NULL,
  most_recent_year = NULL,
  wait_time = 5,
  max_tries = 5
)
}
\arguments{
\item{coord_list}{List of coordinates for which to acquire records}

\item{num_months_backwards}{The number of months backwards from the start
point for which to acquire data. Numeric value. Default is 12. Maximum is 36
(limit imposed by API).}

\item{oldest_month}{Instead of specifying \code{num_months_backwards}, the user
can specify the oldest month of interest (in combination with the oldest year
of interest). Numeric value (e.g., 8 for August).}

\item{oldest_year}{Instead of specifying \code{num_months_backwards}, the user
can specify the oldest year of interest (in combination with the oldest
month of interest). Numeric value (e.g., 2019).}

\item{most_recent_month}{The most recent month of interest. Numeric value
(e.g., 8 for August). By default the function will determine this based on
the most recent API update.}

\item{most_recent_year}{The most recent month of interest. Numeric value
(e.g., 2021). By default the function will determine this based on the most
recent API update.}

\item{wait_time}{If there is a server error when submitting POST request,
the function will wait \code{wait_time} seconds before retrying. Numeric value.
Default is 5.}

\item{max_tries}{Specify the maximum number of times to retry a failed
POST request. Numeric value. Default is 10. Failures are usually due to
timeouts, which are often resolved by retrying, but can also be because there
is no data avaible for the area/date combination.}
}
\value{
A list containing three elements.
\itemize{
\item result: This is the data frame containing the acquired stop records.
Can be assigned as a named data frame using, e.g., if function output has
been called data, data frame can be assigned as \code{df <- data[[1]]}.
\item missing_entries: Data frame which records any local authorities  for which
no data was acquired.
\item server_errors: Data frame which records any unresolved server errors
(i.e., cases where the function stopped trying the POST request).
}
}
\description{
Extracts stop and search records from Police API and organises them by local
authority.
}
\examples{

# get data from the most recently available 12 months (default)
data <- extract_ss_data(coords)

# get data from january 2019 to the most recent date
data <- extract_ss_data(coords, oldest_month = 1, oldest_year = 2019)

# get data between january 2019 and october 2020
data <- extract_ss_data(coords, oldest_month = 1,
                                oldest_year = 2019,
                                most_recent_month = 10,
                                most_recent_year = 2020)

# get data 24 months backwards from august 2021
data <- extract_ss_data(coords, most_recent_month = 8,
                                most_recent_year = 2021,
                                num_months_backwards = 24)

}
