% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newest_data.R
\name{newest_data}
\alias{newest_data}
\title{Retrieve most recent date for which data in API are available}
\usage{
newest_data()
}
\value{
A string describing date in yyyy-mm format
}
\description{
Retrieve most recent date for which data in API are available
}
