% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_data.R
\name{oldest_data}
\alias{oldest_data}
\title{Retrieve oldest date for which data in API are available}
\usage{
oldest_data()
}
\value{
A string describing date in yyyy-mm format
}
\description{
Retrieve oldest date for which data in API are available
}
