% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskratio_from_df.R
\name{riskratio_from_df}
\alias{riskratio_from_df}
\title{Risk ratio from data frame}
\usage{
riskratio_from_df(df, name)
}
\arguments{
\item{df}{Contingency table as data frame on which to calculate risk ratio}

\item{name}{Desired name of resultant data frame containing statistics}
}
\value{
A data frame containing risk ratio and confidence intervals
}
\description{
Calculates relative risk ratio and confidence intervals from a
contingency table constructed for analysis of stop and search disparities
}
\examples{

rr <- riskratio_from_df(df, "Stop and Search disparity")

}
