% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_area_values.R
\name{show_area_values}
\alias{show_area_values}
\title{Show the names of area-based variables}
\usage{
show_area_values(type = NULL)
}
\arguments{
\item{type}{The area type for which to specify names. The options are:
\itemize{
\item la (Local Authority District)
\item county
\item region
\item country
\item force
}

Can also be omitted, in which case all of the above area types will be
returned}
}
\value{
If type is not specified, a data frame containing names for all Local
Authority Districts and their corresponding super-ordinate area data (i.e.,
county, region, country, force).

If type is specified, acharacter vector listing all the names contained
within the area type of interest.
}
\description{
Show all the names of the areas in the built-in coordinate data ('coords').
User can request names for all area types or unique values of a particular
area type. Useful for reference when looking to subset the in-built 'coords'
object.
}
\examples{

# Show me all Local Authority names in 'coords'
show_area_values(type = "la")

# Show me all Forces in 'coords'
show_area_values(type = "force")

# Get all the area types and assign to an object
all_areas <- show_area_values()

}
