% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_coord_list.R
\name{subset_coords}
\alias{subset_coords}
\title{Subset coordinate list}
\usage{
subset_coords(
  coords,
  la = NULL,
  county = NULL,
  region = NULL,
  country = NULL,
  force = NULL
)
}
\arguments{
\item{la}{Character variable or vector of names of Local Authorities to which
to subset.}

\item{county}{Character variable or vector of names of counties to which
to subset.}

\item{region}{Character variable or vector of names of regions to which
to subset.}

\item{country}{Character variable or vector of names of countries to which
to subset.}

\item{force}{Character variable or vector of names of Police Forces to which
to subset.}

\item{coord_list}{The list of coordinates to subset from. In most cases this
will be 'coords', which is bundled with the package.}
}
\value{
A list, which is a subset of 'coord_list', containing the coordinates
and other geographic data for the specified geographies.
}
\description{
Subset the built-in coordinates list to desired areas. User can specify one
or multiple Local Authority, County, Region, Country, or Police Force names.
Note subsetting can only be done by one geography type at a time. Specifying
more than one geography parameter (e.g., la and region) will trigger an error.
}
\examples{

# Subset to just the Metropolitan Police
subset_coord_list <- subset_coords(coords, force = "Metropolitan Police")

# Subset to specific Local Authorities
subset_coord_list <- subset_coords(coords, la = c("Lambeth", "Haringey", "Birmingham"))
# Subset to specific regions
subset_coord_list <- subset_coords(coords, region = c("South East", "South West"))
}
