% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ss_data.R
\name{extract_ss_data}
\alias{extract_ss_data}
\title{Extract stop and search records}
\usage{
extract_ss_data(
  coord_list,
  most_recent_month = NULL,
  most_recent_year = NULL,
  num_months_backwards = 12,
  back_to_month = NULL,
  back_to_year = NULL,
  wait_time = 5,
  max_tries = Inf
)
}
\arguments{
\item{coord_list}{List of coordinates for which to acquire records}

\item{most_recent_month}{The most recent month of interest.
Numeric value (e.g., 8 for August).
If \code{NULL}, the function will determine this based on the most recent API
update.}

\item{most_recent_year}{The most recent month of interest.
Numeric value (e.g., 2021). If \code{NULL}, the function will determine this based
on the most recent API update.}

\item{num_months_backwards}{The number of months backwards from the start
point for which to acquire data. Numeric value. Default is 12. Maximum is 36
(limit imposed by API).}

\item{back_to_month}{Instead of specifying \code{num_months_backwards}, the user
can specify the oldest month of interest (in combination with the oldest year
of interest). Numeric value (e.g., 8 for August).}

\item{back_to_year}{Instead of specifying \code{num_months_backwards}, the user
can specify the oldest month of interest (in combination with the oldest
month of interest). Numeric value (e.g., 2019).}

\item{wait_time}{If there is a server error when submitting POST request,
the function will wait \code{wait_time} seconds before retrying. Numeric value.
Default is 5.}

\item{max_tries}{Specify the maximum number of times to retry a failed
POST request. Numeric value. Default is Infinite, i.e., it will keep trying
until it succeeds. This is because failures are usually due to timeouts that
are resolved by retrying.}
}
\value{
A list containing three elements.
\itemize{
\item result: This is the data frame containing the acquired stop records.
Can be assigned as a named data frame using, e.g., if function output has
been called data, data frame can be assigned as \code{df <- data[[1]]}.
\item missing_entries: Data frame which records any local authorities  for which
no data was acquired.
\item server_errors: Data frame which records any unresolved server errors
(i.e., cases where the function stopped trying the POST request).
}
}
\description{
Extracts stop and search records from Police API and organises them by local
authority.
}
\examples{

data <- extract_ss_data(ordered_coords, back_to_month = 1, back_to_year = 2019)

data <- extract_ss_data(ordered_coords, most_recent_month = 8, most_recent_year = 2021)

}
